Restructure modules and templates, clean up white space of code a little
