+++ portal/templates/layouts/sessions/view_sessions.html
@@ -21,6 +21,9 @@
         <button type="Submit" name="edit" class="course-button">view</button>
       </form>
       </div>
+      <div class="about-course1">
+        <a href="{{ url_for('portal.index') }}" class="course-button">Assignments</a>
+      </div>
   </section>
     {% endfor %}
 </article>
